[lldb-dap] Fix a warning
This patch fixes:

  lldb/tools/lldb-dap/lldb-dap.cpp:1405:16: error: comparison of
  integers of different signs: 'int64_t' (aka 'long') and 'size_type'
  (aka 'unsigned long') [-Werror,-Wsign-compare]

(cherry picked from commit e7a43ca84fabeab386ba4d402167244dac27b265)

+++ lldb/tools/lldb-dap/lldb-dap.cpp
@@ -1403,7 +1403,7 @@ void request_completions(const llvm::json::Object &request) {
   // Handle the offset change introduced by stripping out the
   // `command_escape_prefix`.
   if (had_escape_prefix) {
-    if (offset < g_dap.command_escape_prefix.size()) {
+    if (offset < static_cast<int64_t>(g_dap.command_escape_prefix.size())) {
       body.try_emplace("targets", std::move(targets));
       response.try_emplace("body", std::move(body));
       g_dap.SendJSON(llvm::json::Value(std::move(response)));
